+++ crescent/hooks.py
@@ -1,14 +1,14 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Sequence, TypeVar, Generic, overload, Any
-from crescent.events import EventMeta
+from typing import TYPE_CHECKING, Sequence, TypeVar, Generic, overload, Any, cast
 
-from crescent.internal.app_command import AppCommandMeta
-from crescent.typedefs import EventHookCallbackT
 from hikari import Event
+from crescent.events import EventMeta
+from crescent.internal.app_command import AppCommandMeta
 
 if TYPE_CHECKING:
+    from crescent.typedefs import EventHookCallbackT
     from crescent.internal.includable import Includable
     from crescent.typedefs import HookCallbackT
 
@@ -75,25 +75,11 @@ class _Hook(Generic[IncludableT]):
     def __call__(self, obj: Includable[IncludableT]) -> Includable[IncludableT]:
         if isinstance(obj.metadata, AppCommandMeta):
             obj.metadata.add_hooks(self.callbacks, prepend=True, after=self.after)
-
         elif isinstance(obj.metadata, EventMeta):
-            raise Exception
-
+            # Pyright isn't happy with the type of obj.metadata not being known, but that doesn't matter
+            # for this code to function properly.
+            obj.metadata.add_hooks(self.callbacks, prepend=True, after=self.after)  # pyright: ignore
         else:
             raise TypeError("Unsupported type, this should never happen.")
 
         return obj
-
-def add_hooks(
-    hooks: list[T], after_hooks: list[T], hooks_to_add: Sequence[T], *, prepend: bool, after: bool
-) -> None:
-    def extend_or_prepend(list_to_edit: list[T]) -> None:
-        if prepend:
-            list_to_edit[:0] = hooks_to_add
-        else:
-            list_to_edit.extend(hooks_to_add)
-
-    if not after:
-        extend_or_prepend(hooks)
-    else:
-        extend_or_prepend(after_hooks)
