+++ crescent/utils/__init__.py
@@ -6,5 +6,13 @@ from crescent.utils.any_issubclass import *
 from crescent.utils.gather_iter import *
 from crescent.utils.options import *
 from crescent.utils.tasks import *
+from crescent.utils.hooks import *
 
-__all__: Sequence[str] = ("any_issubclass", "gather_iter", "unwrap", "map_or", "create_task")
+__all__: Sequence[str] = (
+    "add_hooks",
+    "any_issubclass",
+    "gather_iter",
+    "unwrap",
+    "map_or",
+    "create_task",
+)
